requirements.txt
anthropic==0.77.0
backoff==2.2.1
fastapi==0.128.0
json_repair==0.54.2
jsonschema==4.26.0
numpy==2.2.6
openai==2.16.0
opencv-python-headless==4.11.0.86
pillow==12.1.0
playwright==1.44.0
protobuf==6.33.5
pydantic==2.12.5
pytesseract==0.3.13
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
setuptools==78.1.1
tiktoken==0.12.0
toml==0.10.2
uvicorn==0.40.0
python-multipart==0.0.22